Address review on WebChat v2 security-parity tests
Extract the duplicated StubServices facade stub, with_peer helper, and
shared constants from the three audit test files into a single
tests/support/harness.rs pulled in via #[path], and add the coverage the
reviewers flagged: wrong and expired tokens via the ?token= SSE shim are
rejected at the route layer, and a session minted for one tenant fails on
another deployment's store. Make the SSO per-IP rate-limit test's
clone-shared-state assertion explicit and rename the headers-file token
const to ENV_TOKEN for parity. Tests only, no production changes.
